backend/requirements.txt
# backend/requirements.txt

Flask<=2.2.2
flask-cors<=3.0.10
tensorflow==2.18.0
numpy
Pillow<=9.2.0
Werkzeug<=2.2.2